feat: only count authors who have spoken at least 10 tp sentences
this line is negotiable for sure, but is a good starting point to
increase the veracity of the graphs i produce via ilo muni
importantly, it's about as weak a line as it gets; it's very easy to
imagine individuals crossing the 10 line while not actually being toki
pona speakers by accidentally saying one-off matching sentences over the
course of a few months or years. but this removes a huge portion of
errant matches which much more heavily affect authorship presentation
than hits presentation.